Added New Feature For Pickup Items
Pickup items re-spawning now working on server & clients, must be added to the networked pool, once a player pickups the item, all clients will be in sync, no item will be available until its re-spawned, existing pickup items in your project need to be run through the GreedyVox tool to have all the required components.

Networking moving platform is now in synced with server and clients and fixed a buffer overflow bug when platforms spawned into the game world.

Network object pool has been changed to work with UCC, now can have items apart of the pooling system but not counted in the object pool, this way can spawn items over the network that doesn't need to be pooled but needing the pooling system to function correctly with UCC.